Fix MPS compatibility for I2V by adjusting device usage and dtype

+++ wan/configs/wan_i2v_14B.py
@@ -14,7 +14,7 @@ i2v_14B.t5_tokenizer = 'google/umt5-xxl'
 
 # clip
 i2v_14B.clip_model = 'clip_xlm_roberta_vit_h_14'
-i2v_14B.clip_dtype = torch.float16
+i2v_14B.clip_dtype = torch.float32
 i2v_14B.clip_checkpoint = 'models_clip_open-clip-xlm-roberta-large-vit-huge-14.pth'
 i2v_14B.clip_tokenizer = 'xlm-roberta-large'
